[5단계 리팩토링] refac : MVC 패턴 기반으로 리팩토링
