+++ src/main/java/car/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package car.service;
+package car.domain;
 
 public interface MoveStrategy {
     boolean canMove();
