--- src/main/java/car/service/InputType.java
@@ -1,23 +1,0 @@
-package car.service;
-
-public enum InputType {
-    CAR_COUNT("자동차 대수는 몇 대 인가요?", "자동차 대수는 1 이상이어야 합니다."),
-    TRY_COUNT("시도할 회수는 몇 회 인가요?", "시도 회수는 1 이상이어야 합니다."),
-    CAR_NAME("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).", "자동차 이름은 5자를 초과할 수 앖습니다.");
-
-    private final String inputText;
-    private final String rangeErrorMessage;
-
-    InputType(String inputText, String rangeErrorMessage) {
-        this.inputText = inputText;
-        this.rangeErrorMessage = rangeErrorMessage;
-    }
-
-    public String getInputText() {
-        return inputText;
-    }
-
-    public String getRangeErrorMessage() {
-        return rangeErrorMessage;
-    }
-}
