+++ src/main/java/car/view/InputValidator.java
@@ -1,15 +1,19 @@
-package car.service;
+package car.view;
+
+import car.view.InputType;
 
 public class InputValidator {
     public static void validatePositive(int input, InputType inputType) {
-        if (input < 1) {
+        int minValue = inputType.getStandardValue();
+        if (input < minValue) {
             throw new IllegalArgumentException(inputType.getRangeErrorMessage());
         }
     }
 
     public static void validateCarNameLengthInRange(String[] input, InputType inputType) {
+        int maxCarNameLength = inputType.getStandardValue();
         for (String name : input) {
-            if (name.length() > 5) {
+            if (name.length() > maxCarNameLength) {
                 throw new IllegalArgumentException(inputType.getRangeErrorMessage());
             }
         }
