+++ src/test/java/car/CarTest.java
@@ -1,6 +1,6 @@
 package car;
 
-import car.dto.Car;
+import car.domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
