+++ src/test/java/car/InputValidatorTests.java
@@ -1,7 +1,7 @@
 package car;
 
-import car.service.InputType;
-import car.service.InputValidator;
+import car.view.InputType;
+import car.view.InputValidator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
